Integrate peephole optimization into mc-gcc via shared library
You no longer need to run the optimization as a standalone app.

$ make mc-so
$ time ./mc-so tests/sieve.c
$ time ./mc-so -Op tests/sieve.c
$ ./mc-so -Op -o sieve tests/sieve.c
$ time ./sieve

With literally a few more lines of changes, a boot-strapped version
of the compiler can also run the shared object optimizer.  I
didn't add that capability, because I don't like that the shared
object  file is not in a 'well known' library directory, and can
be easily deleted independently from the compiler.